web/requirements.txt
Flask
python-dotenv
Flask-Session
redis
bcrypt
gunicorn